OV-737 : some code improvements
