+++ src/main/kotlin/uk/gov/justice/digital/hmpps/officialvisitsapi/model/response/CreateOfficialVisitResponse.kt
@@ -1,6 +1,7 @@
 package uk.gov.justice.digital.hmpps.officialvisitsapi.model.response
 
 import io.swagger.v3.oas.annotations.media.Schema
+import uk.gov.justice.digital.hmpps.officialvisitsapi.common.VisitorAndContactId
 
 data class CreateOfficialVisitResponse(
   @Schema(description = "The official visit ID")
@@ -9,11 +10,4 @@ data class CreateOfficialVisitResponse(
   val prisonerNumber: String,
   @Schema(description = "The visitor and contact IDs")
   val visitorAndContactIds: List<VisitorAndContactId>,
-)
-
-data class VisitorAndContactId(
-  @Schema(description = "The visitor ID")
-  val visitorId: Long,
-  @Schema(description = "The contact ID (null when no contact was created)")
-  val contactId: Long?,
 )
