+++ src/main/kotlin/uk/gov/justice/digital/hmpps/officialvisitsapi/service/OfficialVisitCancellationService.kt
@@ -4,6 +4,7 @@ import jakarta.persistence.EntityNotFoundException
 import org.slf4j.LoggerFactory
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
+import uk.gov.justice.digital.hmpps.officialvisitsapi.common.VisitorAndContactId
 import uk.gov.justice.digital.hmpps.officialvisitsapi.model.AttendanceType
 import uk.gov.justice.digital.hmpps.officialvisitsapi.model.request.OfficialVisitCancellationRequest
 import uk.gov.justice.digital.hmpps.officialvisitsapi.repository.OfficialVisitRepository
@@ -76,7 +77,7 @@ class OfficialVisitCancellationService(
       officialVisitId = officialVisitId,
       prisonerVisitedId = prisonerVisited.prisonerVisitedId,
       prisonerNumber = prisonerVisited.prisonerNumber,
-      visitorAndContactIds = officialVisit.officialVisitors().map { it.officialVisitorId to it.contactId },
+      visitorAndContactIds = officialVisit.officialVisitors().map { VisitorAndContactId(it.officialVisitorId, it.contactId) },
     ).also {
       logger.info("Official visit with ID $officialVisitId cancelled.")
     }
@@ -87,6 +88,6 @@ class OfficialVisitCancellationService(
     val officialVisitId: Long,
     val prisonerVisitedId: Long,
     val prisonerNumber: String,
-    val visitorAndContactIds: List<Pair<Long, Long?>>,
+    val visitorAndContactIds: List<VisitorAndContactId>,
   )
 }
